AI_Chatbot_for_UNAI/requirements.txt
google-api-python-client==2.187.0
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
pinecone==8.0.0
pinecone-plugin-assistant==3.0.1
python-dotenv==1.0.0
requests==2.32.4
streamlit==1.31.0